model/requirements.txt
django
pillow
transformers
torch
huggingface_hub
